markup & css updates (mobile)
update css to make for better mobile layout, and don't allow article elements to extend outside of the viewport / creating horizontal scrolling.

if the contents of an 'item' extend horizontally beyond the viewport, an overflow: auto; has been set in place, so that item can be scrolled individually of the rest of the document.

Convert article lead-in listings from divs to ul > li > article

add appropriate aria roles and slight modifications to the markup to be more semantically correct.